Fix type of returned value for utility function ToInt

+++ src/function.cpp
@@ -72,16 +72,14 @@ FunctionExtractor::FunctionNameLocation FunctionExtractor::GetNameLocation(const
     std::string coords = function_ast.substr(coord_start + 1, coord_end - coord_start - 1);
 
     size_t comma = coords.find(',');
-    Position start{static_cast<size_t>(ToInt(coords.substr(0, comma))),
-                   static_cast<size_t>(ToInt(coords.substr(comma + 2)))};
+    Position start{ToInt(coords.substr(0, comma)), ToInt(coords.substr(comma + 2))};
 
     size_t dash = function_ast.find('[', coord_end);
     size_t end_bracket = function_ast.find(']', dash);
     std::string end_coords = function_ast.substr(dash + 1, end_bracket - dash - 1);
 
     comma = end_coords.find(',');
-    Position end{static_cast<size_t>(ToInt(end_coords.substr(0, comma))),
-                 static_cast<size_t>(ToInt(end_coords.substr(comma + 2)))};
+    Position end{ToInt(end_coords.substr(0, comma)), ToInt(end_coords.substr(comma + 2))};
 
     return {start, end, ""};
 }
@@ -111,8 +109,7 @@ FunctionExtractor::FindEnclosingClass(const std::string &ast, const FunctionName
         std::string coords = ast.substr(coord_start + 1, coord_end - coord_start - 1);
 
         size_t comma = coords.find(',');
-        Position class_start{static_cast<size_t>(ToInt(coords.substr(0, comma))),
-                             static_cast<size_t>(ToInt(coords.substr(comma + 2)))};
+        Position class_start{ToInt(coords.substr(0, comma)), ToInt(coords.substr(comma + 2))};
 
         size_t dash = ast.find('-', coord_end);
         size_t second_coord_start = ast.find('[', dash);
@@ -120,8 +117,7 @@ FunctionExtractor::FindEnclosingClass(const std::string &ast, const FunctionName
         std::string end_coords = ast.substr(second_coord_start + 1, second_coord_end - second_coord_start - 1);
 
         comma = end_coords.find(',');
-        Position class_end{static_cast<size_t>(ToInt(end_coords.substr(0, comma))),
-                           static_cast<size_t>(ToInt(end_coords.substr(comma + 2)))};
+        Position class_end{ToInt(end_coords.substr(0, comma)), ToInt(end_coords.substr(comma + 2))};
 
         if (func_loc.start.line > class_start.line ||
             (func_loc.start.line == class_start.line && func_loc.start.col >= class_start.col)) {
